cgai_backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
firebase-admin==7.1.0
django-cors-headers==4.9.0
python-dotenv==1.1.1
psycopg2-binary==2.9.11